moved all non-BasicParameters into own field
